requirements.txt
boto3==1.34.131
requests==2.31.0
python-dotenv==1.0.0